feat(orga): enhance stars display for average stage
