+++ orga/app/components/campaign/cards/stage-average.hbs
@@ -6,5 +6,5 @@
   @isLoading={{@isLoading}}
   ...attributes
 >
-  <Campaign::StageStars @result={{this.valuePercentage}} @stages={{@stages}} @isWide={{true}} />
+  <Campaign::StageStars @totalStage={{@totalStage}} @reachedStage={{@reachedStage}} @isWide={{true}} />
 </Ui::IndicatorCard>
