--- orga/app/components/campaign/cards/stage-average.js
@@ -1,7 +1,0 @@
-import Component from '@glimmer/component';
-
-export default class StageAverage extends Component {
-  get valuePercentage() {
-    return Math.round(this.args.value * 100);
-  }
-}
