+++ orga/app/components/campaign/results/assessment-row.hbs
@@ -14,7 +14,7 @@
   {{/if}}
   {{#if @hasStages}}
     <td>
-      <Campaign::StageStars
+      <Campaign::StageStarsOld
         @result={{this.valuePercentage}}
         @stages={{@stages}}
         @withTooltip={{true}}
