+++ orga/app/components/campaign/stage-stars.hbs
@@ -6,24 +6,4 @@
     @color="blue"
     class="stage-stars {{if @isWide 'stage-stars--wide'}}"
   />
-  {{#if this.displayTooltip}}
-    <PixTooltip
-      @id="stage-tooltip-{{this.reachedStage.id}}"
-      @position={{@tooltipPosition}}
-      @isWide={{true}}
-      class="hide-on-mobile"
-    >
-      <:triggerElement>
-        <FaIcon
-          @icon="circle-info"
-          class="stage-stars__info-icon"
-          tabindex="0"
-          aria-describedby="stage-tooltip-{{this.reachedStage.id}}"
-        />
-      </:triggerElement>
-      <:tooltip>
-        {{this.tooltipText}}
-      </:tooltip>
-    </PixTooltip>
-  {{/if}}
 </div>
